Move frontend tests to __test__ folder
